GIS_Integration/requirements.txt
fastapi[all]
uvicorn
requests
python-dotenv
Pillow
boxsdk[jwt]
pyproj
